requirements.txt
Flask==3.1.3
openai
langchain
langchain-community
langchain-huggingface
langchain-text-splitters
sentence-transformers
python-docx
supabase
pgvector
